Streamlit_Version/requirements.txt
streamlit
langchain
faiss-cpu
sentence-transformers
pypdf2
python-docx
pdfplumber